Final legend tweaks
