+++ src/styles/legend-button.scss
@@ -1,7 +1,7 @@
 @import 'styles/settings';
 
 .legendButton {
-  height: 26px;
+  height: 14px;
   width: 26px;
   padding: 0;
   z-index: 2;
